Moved all src files into the src directory
Also fixed the "isNetworkConnected" for ESP8266
